fewfwefwefwefwef

+++ scripts/talker.py
@@ -225,7 +225,7 @@ class ManySyncListener:
 
         arr = arr.reshape(6,-1).T # (N, 6)
         xyz = arr[:,:3]
-        colors = arr[:,3:].astype(np.uint8)
+        colors = arr[:,3:].astype(np.uint32)
         colors = colors[:,0] * BIT_MOVE_16 +colors[:,1] * BIT_MOVE_8 + colors[:,2]
         cloud_data=rfn.merge_arrays(xyz,colors)
         # arr = np.rec.fromarrays((arr[0],arr[1],arr[2],arr[3].astype(np.uint8),arr[4].astype(np.uint8),arr[5].astype(np.uint8)))
